+++ include/robot.h
@@ -12,8 +12,8 @@ class Controller;
 
 class Robot {
 	// Robot Data
-	int current_battery;
-	int max_battery;
+	float current_battery;
+	float max_battery;
 	int current_steps;
 	int max_steps;
 	Controller* controller;
@@ -27,12 +27,12 @@ class Robot {
 	// Helper Methods
 	int calculate_dirt() const;
 public:
-	Robot(house& model, int max_battery, int max_steps, int starting_row, int starting_col);
+	Robot(house& model, float max_battery, int max_steps, int starting_row, int starting_col);
 	~Robot();
 	int clean_house(std::ofstream&);
 	int get_dirt_underneath() const;
 	bool is_wall(Direction) const;
-	int remaining_battery() const;
+	float remaining_battery() const;
 };
 
 #endif
